move republisher into its own package
