+++ pkg/util/nativeutils/republisher/republisher_test.go
@@ -1,13 +1,12 @@
-package agreement_test
+package republisher_test
 
 import (
 	"bytes"
 	"testing"
 
-	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus/agreement"
 	"github.com/dusk-network/dusk-blockchain/pkg/p2p/wire/topics"
 	"github.com/dusk-network/dusk-blockchain/pkg/util/nativeutils/eventbus"
-	"github.com/dusk-network/dusk-wallet/key"
+	"github.com/dusk-network/dusk-blockchain/pkg/util/nativeutils/republisher"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -16,8 +15,8 @@ func TestRepublisher(t *testing.T) {
 	gossipChan := make(chan bytes.Buffer, 1)
 	gl := eventbus.NewChanListener(gossipChan)
 	eb.Subscribe(topics.Gossip, gl)
-	k, _ := key.NewRandConsensusKeys()
-	agreement.NewFactory(eb, k)
+
+	republisher.New(eb, topics.Agreement)
 
 	mockAggro := bytes.NewBuffer([]byte{1})
 	eb.Publish(topics.Agreement, mockAggro)
